Revert "move dropdown into ui-components"
This reverts commit 13d787befafd70bee0213453a34048776330b0f2.
