+++ packages/editor-common/web/statics/styles/dropdown.scss
