+++ packages/ui-components/web/src/index.ts
@@ -22,8 +22,10 @@ export { default as SliderWithInput } from './Components/SliderWithInput';
 export { default as Slider } from './Components/Slider';
 export { default as TextInput } from './Components/TextInput';
 export { Tab, Tabs } from './Components/Tabs';
+// export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
 export { default as Loader } from './Components/Loader';
 export { default as MediaItemErrorMsg } from './Components/MediaItemErrorMsg';
 export { default as TextSearchInput } from './Components/TextSearchInput';
 export { default as Separator } from './Components/Separator';
-export { default as Dropdown } from './Components/Dropdown';
+
+// export { getSelectionStyles } from './Utils/inlineStyleUtils';
